refactor: extract ResidentUI
