+++ src/main/java/com/holub/life/Resident.java
@@ -3,6 +3,7 @@ package com.holub.life;
 import com.holub.model.Cell;
 import com.holub.model.DummyCell;
 import com.holub.ui.CellUI;
+import com.holub.ui.ResidentUI;
 import java.awt.*;
 import com.holub.ui.Colors;  // Contains constants specifying various
 // colors not defined in java.awt.Color.
@@ -15,15 +16,20 @@ import com.holub.ui.Colors;  // Contains constants specifying various
  * @include /etc/license.txt
  */
 
-public final class Resident implements Cell, CellUI {
+public final class Resident implements Cell {
+  ResidentUI residentUI;
+  public Resident() {
+    super();
+    residentUI = new ResidentUI(this);
+  }
   @Override
   public CellUI getCellUI() {
-    return this;
+    return residentUI;
   }
 
-  private static final Color BORDER_COLOR = Colors.DARK_YELLOW;
-  private static final Color LIVE_COLOR = Color.RED;
-  private static final Color DEAD_COLOR = Colors.LIGHT_YELLOW;
+  public void toggle() {
+    amAlive = !amAlive;
+  }
 
   private boolean amAlive = false;
   private boolean willBeAlive = false;
@@ -104,26 +110,6 @@ public final class Resident implements Cell, CellUI {
     amAlive = willBeAlive;
     return changed;
   }
-
-  public void redraw(Graphics g, Rectangle here, boolean drawAll) {
-    g = g.create();
-    g.setColor(amAlive ? LIVE_COLOR : DEAD_COLOR);
-    g.fillRect(here.x + 1, here.y + 1, here.width - 1, here.height - 1);
-
-    // Doesn't draw a line on the far right and bottom of the
-    // grid, but that's life, so to speak. It's not worth the
-    // code for the special case.
-
-    g.setColor(BORDER_COLOR);
-    g.drawLine(here.x, here.y, here.x, here.y + here.height);
-    g.drawLine(here.x, here.y, here.x + here.width, here.y);
-    g.dispose();
-  }
-
-  public void userClicked(Point here, Rectangle surface) {
-    amAlive = !amAlive;
-  }
-
   public void clear() {
     amAlive = willBeAlive = false;
   }
